Move logging to common package so it can be reused
