+++ http/service.go
@@ -20,7 +20,6 @@ import (
 )
 
 const (
-	DefaultLogLevel      = LogDebug
 	DefaultListenAddress = "127.0.0.1:8080"
 )
